% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctmatch.R
\name{correctly_matched}
\alias{correctly_matched}
\title{Calculates the percentage of the population of the second shapefile correctly matched by knowing only the first level}
\usage{
correctly_matched(overlap_output, id1, id2)
}
\arguments{
\item{overlap_output}{The cleaned output from the overlapfunction.}

\item{id1}{The ID field of the first shapefile of interest}

\item{id2}{The ID field of the second shapefile of interest}
}
\value{
The best match weighted population (i.e. how nested) for the second level of geography. The data frame values are as 
follows:
\itemize{
   \item id = The user specified ID for the second level of geography. 
   \item correctly_matched_pop = The weighted and summed population of the first level of geography with the second level of geography that
   it shares the most overlap with. 
   \item pop_wt = The calculated total population of the second level of geography
   \item correct_pct = The percentage of the population that would be correctly matched to the second level of geography knowing only 
   the information from the first level of geography as opposed to some more specific spatial information, such as points/addresses. 

}
}
\description{
This function weights the population of the first level of geography by the proportion of the population with the second level of
geography that it shares the most overlap with. The data is then collapsed and divided by the population of the second level of 
geography in order to determine the population of the second level of geography that could be reached out to successfully by 
knowing only the first level of geography as opposed to point data.
}
\examples{
correctly_matched
}
