% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapfunction.R.R
\name{overlapfunction}
\alias{overlapfunction}
\title{Calculate dyadic population overlap between 2 levels of geography}
\usage{
overlapfunction(filename1a, id1, id2, pop_field, merge_table, merge_id1,
  merge_id2, census_fields)
}
\arguments{
\item{filename1a}{The DBF file name or object of the raw intersectional geographic shapefile}

\item{id1}{The ID field of the first shapefile of interest}

\item{id2}{The ID field of the second shapefile of interest}

\item{pop_field}{The field with the population data}

\item{merge_table}{An optional command to merge on additional census data for sociodemographic information}

\item{merge_id1}{Given the choice to merge on additional data, the name of the field from the first dbf by which to merge}

\item{merge_id2}{Given the choice to merge on additional data, the name of the field from the second dbf by which to merge}

\item{census_fields}{The option to provide 2 or more census fields to calculate the demographics of from Census data}
}
\value{
The data frame with the dyadic population overlap between the first and second shapefiles. The following are the values:
   \itemize{
   \item id2 = The specified ID of the second shapefile set by the user. 
   \item id1 = The specified ID of the first shapefile set by the user. 
   \item DYAD_ID = The pasted and underscored dyadic ID of the two shapefiles 
   \item pop_wt = The Calculated population within the dyad 
   \item first_shp_pop = The calculated total population of the first shapefile
   \item second_shp_pop = The calculated total population of the second shapefile 
   \item overlap1 = A score between 0 and 1 of the population overlap between the first and second shapefiles. Score of 1 = complete nesting
   \item overlap2 = A score between 0 and 1 of the proportion of the second shapefile's population present within the first shapefile. 
   \item optional returns: (census_field)_i_wt = The calculated weighted census field population within the dyad
   \item shp1_(census_field)_i_wt = The calculated weighted census field population within the first level of geography
   \item shp2_(census_field)_i_wt = The calculated weighted census field population within the second level of geography
   \item (census_field)_i_pct_dyad = The calculated weighted census field percent of population within the dyad
   \item (census_field)_i_pct_shp1 = The calculated weighted census field percent of population within the first level of geography
   \item (census_field)_i_pct_shp2 = The calculated weighted census field percent of population within the second level of geography 
}
}
\description{
This function takes the raw output of the three way intersection between 2 levels of geography and census tabulation units in order 
to calculate the population overlap between the 2 levels of geography.
}
\examples{
setwd(wd_zip)
zip_cd115V<-read.dbf("all_merge3.dbf") #481946 obs, 76 vars 
zip_cd115V <- subset(zip_cd115V, POP2010 >= 0)
overlap115 <- overlapfunction(zip_cd115V, id1 =  "ZCTA5CE10", id2="GEOID", pop_field = "POP2010")
overlap115b <- overlapfunction(zip_cd115V, id1 =  "ZCTA5CE10", id2="GEOID", pop_field = "POP2010", census_fields = c("WHITE","HISPANIC") )
}
